requirements.txt
black==22.1.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
pathspec==0.9.0
platformdirs==2.5.1
protobuf==3.19.4
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.11.2
pymongo==4.0.2
python-dotenv==0.19.2
rich==11.2.0
snakeviz==2.1.1
tabulate==0.8.9
tomli==2.0.1
tornado==6.1
tqdm==4.63.0
